Clean up parsers, add documentURL to syntax exceptions
