+++ src/main/scala/wdlTools/util/SourceCode.scala
@@ -23,25 +23,23 @@ object SourceCode {
   //   foo.txt
   //
   // Follow the URL and retrieve the content as a string.
-  case class Loader(opts: Options) {
-    private def fetchHttpAddress(url: URL): Seq[String] = {
-      val src = Source.fromURL(url)
-      try {
-        src.getLines().toVector
-      } finally {
-        src.close()
-      }
+  private def fetchHttpAddress(url: URL): Seq[String] = {
+    val src = Source.fromURL(url)
+    try {
+      src.getLines().toVector
+    } finally {
+      src.close()
+    }
+  }
+
+  def loadFrom(url: URL): SourceCode = {
+    val lines = url.getProtocol match {
+      case "http"  => fetchHttpAddress(url)
+      case "https" => fetchHttpAddress(url)
+      case "file"  => Util.readLinesFromFile(Paths.get(url.getPath))
+      case _       => throw new Exception(s"unknown protocol in URL ${url}")
     }
 
-    def apply(url: URL): SourceCode = {
-      val lines = url.getProtocol match {
-        case "http"  => fetchHttpAddress(url)
-        case "https" => fetchHttpAddress(url)
-        case "file"  => Util.readLinesFromFile(Paths.get(url.getPath))
-        case _       => throw new Exception(s"unknown protocol in URL ${url}")
-      }
-
-      SourceCode(url, lines)
-    }
+    SourceCode(url, lines)
   }
 }
